add constant for button spacing; seperate out test files
